pyproject.toml
[project]
name = "ir-insect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "picamera2>=0.3.30",
    "rpi-gpio>=0.7.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.9",
]
